refactor: rename utlis to util
