+++ weasel/cli/dvc.py
@@ -6,8 +6,8 @@ from typing import Any, Dict, List, Optional
 
 from wasabi import msg
 
-from weasel.utils import get_hash, join_command, load_project_config, run_command
-from weasel.utils import working_dir
+from weasel.util import get_hash, join_command, load_project_config, run_command
+from weasel.util import working_dir
 
 from .main import COMMAND, NAME, PROJECT_FILE, Arg, Opt, app
