+++ weasel/cli/run.py
@@ -8,10 +8,10 @@ import typer
 from wasabi import msg
 from wasabi.util import locale_escape
 
-from weasel.utils import SimpleFrozenDict, SimpleFrozenList, check_spacy_env_vars
-from weasel.utils import get_checksum, get_hash, is_cwd, join_command
-from weasel.utils import load_project_config, parse_config_overrides, run_command
-from weasel.utils import split_command, working_dir
+from weasel.util import SimpleFrozenDict, SimpleFrozenList, check_spacy_env_vars
+from weasel.util import get_checksum, get_hash, is_cwd, join_command
+from weasel.util import load_project_config, parse_config_overrides, run_command
+from weasel.util import split_command, working_dir
 
 from .main import COMMAND, PROJECT_FILE, PROJECT_LOCK, Arg, Opt, app
