+++ weasel/util/__init__.py
