+++ weasel/util/commands/hashing.py
