+++ weasel/util/config.py
