+++ weasel/util/environment.py
