+++ weasel/util/filesystem.py
