+++ weasel/util/modules.py
