+++ weasel/util/structures.py
